+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.2.7] - 2023-02-08
+
+### Updated
+
+- Security patches for npm and pip packages
+- Added unit tests for JS Lambda Hook SDK
+
 ## [5.2.6] - 2023-01-11
 
 ### Updated